requirements.txt
crewai==0.157.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
